Add new folder for templates.
Add new folder for templates and a new subfolder for AWS CloudFormation
templates.

Add new CloudFormation template for testing X-Road REST Adapter Service
in AWS. The template creates a VPC and adds an EC2 instance with an
Elastic IP address and a security group. After that the latest version
of REST Adapter Service is fetched from GitHub, compiled and started.